build: move swagger files
