+++ tdm-be/src/app.ts
@@ -9,7 +9,7 @@ const dataEnrichmentRoute = require('./controller/data-enrichment');
 const dataWrapperRoute = require('./controller/data-wrapper');
 const traitmentRoute = require('./controller/traitment');
 const webhookRoute = require('./controller/webhook');
-const swaggerFile = require('../swagger-config.json');
+const swaggerFile = require('./swagger/swagger-config.json');
 const cors = require('cors');
 const basicAuth = require('express-basic-auth'); // This package is used for basic authentication
 const cron = require('node-cron');
