+++ tdm-be/tsconfig.json
@@ -11,10 +11,12 @@
         "skipLibCheck": true,
         "forceConsistentCasingInFileNames": true,
         "resolveJsonModule": true,
-        "allowSyntheticDefaultImports": true
+        "allowSyntheticDefaultImports": true,
+        "allowJs": true
     },
     "include": [
         "src/**/*.ts",
+        "src/**/*.js",
         "config/config.json"
     ],
     "exclude": [
